requirements.txt
itsdangerous
flask
flask-Restful
pyjwt
flask-sqlalchemy
werkzeug
flask-mail
pymysql
mysql-client
